% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_rpt.R
\name{crt_rpt}
\alias{crt_rpt}
\title{Title Create Sassy report object by using titles and footnotes from DLV}
\usage{
crt_rpt(
  env_list,
  output_name,
  output_type,
  headers_list,
  footers_vector,
  tbl,
  margin_list = NULL
)
}
\arguments{
\item{env_list}{An object that must contain study root, program name, program side.}

\item{output_name}{TLF output name.}

\item{output_type}{Output file type. Refer to output_type in create_report function in reporter.}

\item{headers_list}{Page titles and table titles.}

\item{footers_vector}{Page footers and table footers.}

\item{tbl}{Table data.}

\item{margin_list}{Margins for the report (top, bottom, left, right).}
}
\value{
An report object.
}
\description{
Title Create Sassy report object by using titles and footnotes from DLV
}
