% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlv.R
\name{file_name_repair}
\alias{file_name_repair}
\title{Handling of column names. Ensures column names are not empty and are unique.}
\usage{
file_name_repair(nms)
}
\arguments{
\item{nms}{Column Names}
}
\value{
List of validated column names
}
\description{
Handling of column names. Ensures column names are not empty and are unique.
}
