% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footers_from_dlv.R
\name{footers_from_dlv}
\alias{footers_from_dlv}
\title{Title read DLV file and format footnotes.}
\usage{
footers_from_dlv(env_list, output_name, dlv_full_path = NULL)
}
\arguments{
\item{env_list}{An object that must contain study root, program name, program side.}

\item{output_name}{TLF output name.}

\item{dlv_full_path}{Assigned DLV full path, including file name.}
}
\value{
A list contains page footnotes.
}
\description{
Title read DLV file and format footnotes.
}
