% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers_from_dlv.R
\name{headers_from_dlv}
\alias{headers_from_dlv}
\title{Title To read DLV file and format page header and titles.}
\usage{
headers_from_dlv(env_list, output_name, dlv_full_path = NULL)
}
\arguments{
\item{env_list}{An object that must contain study root, program name, program side.}

\item{output_name}{TLF output name.}

\item{dlv_full_path}{Assigned DLV full path, including file name.}
}
\value{
A list contains page header and titles.
}
\description{
Title To read DLV file and format page header and titles.
}
