% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlv.R
\name{read_dlv}
\alias{read_dlv}
\title{Read Deliverable file}
\usage{
read_dlv(
  study_root = NULL,
  pgm_name = NULL,
  pgm_side = NULL,
  output_name = NULL,
  dlv_full_path = NULL
)
}
\arguments{
\item{study_root}{Study root path.}

\item{pgm_name}{Program short name (without file extension name)}

\item{pgm_side}{Program Side (prod / qc)}

\item{output_name}{Specific output name. If null, return all the records.}

\item{dlv_full_path}{DLV file full path, including the file name and extension file name.}
}
\value{
A list includes global variables, outputs, libraries, and deliverable file path
}
\description{
Read Deliverable file
}
